requirements.txt
django==2.2.12
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
pymongo==3.10.0